WIP: Add building inequality constraint vector h
For the condensed problem, we have the inequality constraints:
P*z <= h

h and P must be constructed from the prb data. We have added
a PoC of how to do it, but it has not been tested thoroughly.
TODO: the problem is that P has a couple of zero rows, due to
the parameter x_0. This makes the PBM fail to find an interior
point. The data generation must eliminate all constraints that
the optimization variable can not influence/enforce
(like the constraints on the initial state x_0).